Remove unnecessary component layer
