+++ src/components/thresholds/ThresholdsOverview.vue
@@ -1,46 +1,90 @@
 <template>
-  <SidePanelContent type="thresholds" title="Thresholds">
-    <ThresholdsPanel
-      v-model:selectedWarningLevelIds="
-        warningLevelsStore.selectedWarningLevelIds
-      "
-      :warningLevels="warningLevelsStore.warningLevels"
-      :crossings="warningLevelsStore.thresholdCrossings"
-      :selectedThresholdCrossings="
-        warningLevelsStore.selectedThresholdCrossings
-      "
-      :locationIds="props.locationIds"
-      :selectable="nodeCanShowThresholds"
-      @close="sidePanelStore.toggleActive('thresholds')"
-      @navigate="emit('navigate', $event)"
-    />
-  </SidePanelContent>
+  <div>
+    <v-chip-group
+      class="px-2 py-2 d-flex flex-wrap flex-0-0"
+      v-model="warningLevelsStore.selectedWarningLevelIds"
+      multiple
+      column
+      selected-class="v-chip--variant-tonal"
+    >
+      <v-chip
+        v-for="level in warningLevelsStore.warningLevels"
+        :key="level.id"
+        :value="level.id"
+        :text="level.name"
+        label
+        variant="text"
+        class="pe-0 ps-2"
+        border
+      >
+        <template #prepend>
+          <v-img width="20" height="20" :src="level.icon" class="me-1" />
+        </template>
+        <template #append>
+          <v-chip
+            :text="level.count"
+            density="compact"
+            variant="flat"
+            class="ms-2 pa-1 pointer-events-none"
+            size="small"
+          />
+        </template>
+      </v-chip>
+    </v-chip-group>
+    <div v-if="warningLevelsStore.warningLevels.length === 0" class="pa-2">
+      {{ t('thresholds.noThresholdCrossing') }}
+    </div>
+    <!-- Important to have item-height as it greatly improves performance -->
+    <v-virtual-scroll
+      ref="virtualScroll"
+      class="scroll-container flex-1-1"
+      :items="groupedCrossings"
+      :item-height="52"
+    >
+      <template #default="{ item: crossingsGroup }">
+        <div class="my-1 mx-2">
+          <ThresholdSummary
+            :crossings="crossingsGroup"
+            :isSelected="crossingsGroup[0].locationId === locationIds"
+            :selectable="selectable"
+            @navigate="emit('navigate', $event)"
+          />
+        </div>
+      </template>
+    </v-virtual-scroll>
+  </div>
 </template>
+
 <script setup lang="ts">
-import SidePanelContent from '@/components/sidepanel/SidePanelContent.vue'
-import { ref, watch } from 'vue'
-import ThresholdsPanel from '@/components/thresholds/ThresholdsPanel.vue'
-import type { TopologyNode } from '@deltares/fews-pi-requests'
-import { useWarningLevelsStore } from '@/stores/warningLevels'
-import { useSidePanelStore } from '@/stores/sidePanel'
+import {
+  LevelThresholdCrossings,
+  TopologyNode,
+} from '@deltares/fews-pi-requests'
+import { computed, watch, nextTick, onMounted, useTemplateRef } from 'vue'
+import { useI18n } from 'vue-i18n'
+
 import { NavigateRoute } from '@/lib/router/types'
+
+import { useWarningLevelsStore } from '@/stores/warningLevels'
+
+import ThresholdSummary from '@/components/thresholds/ThresholdSummary.vue'
 import { nodeHasMap } from '@/lib/topology/nodes'
+
+const { t } = useI18n()
+const warningLevelsStore = useWarningLevelsStore()
 
 interface Props {
   topologyNode?: TopologyNode
   locationIds?: string
 }
 
-interface Emits {
-  navigate: [to: NavigateRoute]
-}
-const emit = defineEmits<Emits>()
-
 const props = defineProps<Props>()
 
-const sidePanelStore = useSidePanelStore()
-const warningLevelsStore = useWarningLevelsStore()
-const nodeCanShowThresholds = ref(false)
+interface Emits {
+  navigate: [to: NavigateRoute]
+  close: []
+}
+const emit = defineEmits<Emits>()
 
 watch(
   () => props.topologyNode?.id,
@@ -50,27 +94,72 @@ watch(
   },
 )
 
-watch(
-  () => props.topologyNode,
-  (newNode) => {
-    if (newNode && nodeHasMap(newNode)) {
-      nodeCanShowThresholds.value = true
-    } else {
-      nodeCanShowThresholds.value = false
+const selectable = computed<boolean>(() => {
+  const node = props.topologyNode
+  return node !== undefined && nodeHasMap(node)
+})
+
+const virtualScroll = useTemplateRef('virtualScroll')
+
+const groupedCrossings = computed(() => {
+  const grouped: Record<string, LevelThresholdCrossings[]> = {}
+  warningLevelsStore.selectedThresholdCrossings.forEach((crossing) => {
+    const key = crossing.locationId
+    if (!grouped[key]) {
+      grouped[key] = []
     }
-  },
-  { immediate: true },
-)
+    grouped[key].push(crossing)
+  })
+  return Object.values(grouped)
+})
+
+// Scroll to the selected item
+async function scrollToSelectedItem() {
+  if (!virtualScroll.value) return
+  const selectedIndex = groupedCrossings.value.findIndex(
+    (group) => group[0].locationId === props.locationIds,
+  )
+  if (selectedIndex !== -1) {
+    // Wait for the next DOM update cycle
+    await nextTick()
+    // Scroll to the item
+    virtualScroll.value.scrollToIndex(selectedIndex)
+  }
+}
+
+watch([() => props.locationIds, groupedCrossings], () => {
+  scrollToSelectedItem()
+})
+
+// Scroll to the selected item when we directly navigate to the page
+onMounted(() => {
+  scrollToSelectedItem()
+})
 </script>
 
 <style scoped>
-.refresh-container {
-  height: 28px;
+.threshold-summary-container {
+  width: 450px;
 }
 
-.task-runs-panel {
-  width: 450px;
-  display: grid;
-  grid-template-rows: auto 1fr auto auto;
+.warning-count {
+  text-align: center;
+}
+
+.v-list-item-subtitle {
+  font-size: 0.8em;
+  color: var(--v-theme-on-surface);
+  opacity: 1;
+  width: 100%;
+  text-align: center;
+}
+
+:deep(.v-list-item__content) {
+  overflow: visible !important;
+  width: 65px;
+}
+
+:deep(.v-chip__content) {
+  width: 100%;
 }
 </style>
